refactor(lowy): move scroll-request state to its own module
`useCommentScrollRequest` and `ScrollRequest` lived next to the
composer state machine but share nothing with it — the composer
tracks an in-flight draft; the scroll request is a one-shot
"navigation lands → scroll to this anchor" hand-off.

Move both to `scrollRequest.ts`. Composer state stays focused on
the draft machine; the new file is small but has a single concern
(volatility-aligned per Lowy).
